refactor(storage): inline object lock metadata writes
